TESTNG-459: enable 'Convert to YAML' context menu only for suite files in XML format
Commit 6461478460cc7bac13e7 introduced new property tester variable
isXmlSuite which is used to offer the 'Convert to YAML' menu only on
XML files containing '<suite>' tag.

Commit 7136722da8218d7a2e3d added context menu to convert .xml suites
to .yaml. But it was enabled for all XML files regardless of their
actual content.

Enable 'Convert to YAML' only for files that recognized as XML suites
by property tester variable isXmlSuite.

Signed-off-by: Mykola Nikishov <mn@mn.com.ua>