Better tab handling.
Tabs used to be defined in plug-in.xml via a .xsd extension
but I think that's overkill. Removed this mention in the plugin.xml
and initializing the tabs directly in TestRunnerViewPart.

Also split the tabs between those that report right away
(listening tabs) and those that report at the end of the
test run (reporting tabs).