This commit fixes an issue with TestNG groups run configuration.
Problem: a user creates a new TestNG run configuration and selects some groups to run. The user runs the configuration and opens it again
and press the "Run" button, but no tests are run.

Solution: all data from m_groupMap of GroupSelector has been copied to valueMap of MultiSelector in the initializeFrom method.